requirements.txt
pandas==2.0.2
Pillow==9.2.0
Pillow==9.5.0
Requests==2.31.0
